fix: 清理 clippy warnings (15→1)
- derivable_impls: LabelFieldType, StoreInfo, ShiftStatus 改用 #[derive(Default)]
- match_like_matches_macro: quick_tcp_probe 改用 matches!()
- collapsible_if: event_router, employee repo, cancel_payment, archive_worker
- unnecessary_cast: 移除 3 处 `quantity as i32` 冗余转换

剩余 1 warning: jwt.rs too_many_arguments (需重构 API 签名)

+++ edge-server/src/orders/appliers/order_completed.rs
@@ -29,7 +29,7 @@ impl EventApplier for OrderCompletedApplier {
             for item in &snapshot.items {
                 snapshot
                     .paid_item_quantities
-                    .insert(item.instance_id.clone(), item.quantity as i32);
+                    .insert(item.instance_id.clone(), item.quantity);
             }
 
             // Update sequence and timestamp
